refactor(itemmanager): 重构物品管理器
- 将 ItemSource 中的静态方法移动到 ItemManager 类中懒加载
- 优化了物品源的注册和获取逻辑
- 调整了 parse2ItemStack 方法的实现
- 移除了 ItemSource 中的冗余代码
